+++ src/main/java/packages/query/RangeQuery.java
@@ -12,18 +12,18 @@ import packages.userinteraction.Mapping;;
 public class RangeQuery {
 
     Mtree root;
-    Double searchRadius;
+    Float searchRadius;
     String query;
     String question;
-    public Map<String,Double> relevantQuestions;
+    public Map<String,Float> relevantQuestions;
     Tfidf tfidf;
     Mapping mapping;
-    Double socialDist;
-    public RangeQuery(Mtree r, Double search, String Query, String question,Tfidf tfidf_, Mapping m,Double SocialDist) {
+    Float socialDist;
+    public RangeQuery(Mtree r, Float search, String Query, String question,Tfidf tfidf_, Mapping m,Float SocialDist) {
         this.root = r;
         this.searchRadius = search;
         this.query = Query;
-        this.relevantQuestions = new HashMap<String,Double>();
+        this.relevantQuestions = new HashMap<String,Float>();
         this.tfidf = tfidf_;
         this.socialDist = SocialDist;
         this.mapping = m;
@@ -31,7 +31,7 @@ public class RangeQuery {
         load();
     }
 
-    public Map<String,Double> getRelevantQids() {
+    public Map<String,Float> getRelevantQids() {
         return relevantQuestions;
     }
 
@@ -52,13 +52,13 @@ public class RangeQuery {
 
             //System.out.println("\n____________");
             Mtree node = queue.remove();
-            Double queryDistance = tfidf.getquerydistance(node.getQid(), query);
+            Float queryDistance = tfidf.getquerydistance(node.getQid(), query);
             if (node.isLeaf()) {
                 //System.out.println(" Leaf Node : " + node.getQid() + " Querydistance : " + queryDistance + "  : " );
 
                 for (Mtree child : node.getChildren()) {
                     if (queryDistance - tfidf.getdistance(node.getQid(), child.getQid()) < searchRadius) {
-                        Double dist = tfidf.getquerydistance(child.getQid(), query);
+                        Float dist = tfidf.getquerydistance(child.getQid(), query);
                         if (dist < searchRadius && mapping.getSocialDistance(child.getQid(), question )<socialDist) {
                             relevantQuestions.put(child.getQid(),dist);
                         }
@@ -70,7 +70,7 @@ public class RangeQuery {
                     //System.out.println(" query dist :  "+queryDistance + " Parent dist : " + tfidf.getdistance(node.getQid(), child.getQid() )+ " Noderadius : "+ node.getRadius());
                     if (queryDistance - tfidf.getdistance(node.getQid(), child.getQid()) < searchRadius
                             + node.getRadius()) {
-                        Double dist = tfidf.getquerydistance(child.getQid(), query);
+                        Float dist = tfidf.getquerydistance(child.getQid(), query);
                         if (dist < searchRadius + node.getRadius() && mapping.getSocialDistance(child.getQid(), question )<socialDist) {
                             queue.add(child);
                         }
